自动格式化json字符串
