+++ views/guide.hbs
@@ -1,0 +1,17 @@
+<h2>模块页面</h2>
+<h3>登陆接口配置</h3>
+<h3>测试接口配置</h3>
+<h3>排序</h3>
+<h3>查看参数</h3>
+<h3>查看接口</h3>
+<h3>接口测试</h3>
+
+<h2>接口编辑页面</h2>
+<h3>mockjs</h3>
+<h3>json schema</h3>
+<h3>复制接口</h3>
+
+<h2>账号操作</h2>
+
+<h2>接口升级</h2>
+node bin/www --update --hdjk
