+++ views/readme.hbs
@@ -3,6 +3,16 @@
   校验规则示例：
 </p>
 <pre>
+/**
+ * 根元素是一个json对象，有下列属性："name","_BIZCODE","age","list"，其中"_BIZCODE"是必要属性，其它属性可选
+ * "name"属性值是字符串，表示姓名
+ * "_BIZCODE"属性值是字符串，值只能为"0002"
+ * "age"属性是值是整数，表示年龄，值只能在18~50之间
+ * "list"属性值是数组，数组的元素为json对象，该数组最小长度为1，
+ * 每个数组元素的json对象都必须包含"id"，"name"属性，
+ * "id"是的值是字符串，必须为"a"或者"b"
+ * "name"的值是字符串，必须匹配"[/w]{2,4}"的正则表达式且长度在2~4。
+ */
 {
 	"type": "object",
 	"properties": {
